requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.3
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-filter==24.3
django-phonenumber-field==8.0.0
django-resized==1.0.3
djangorestframework==3.15.2
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
packaging==24.2
phonenumberslite==8.13.52
pillow==11.1.0
pyaes==1.6.1
pyasn1==0.6.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9
sqlparse==0.5.2
Telethon==1.38.1
uritemplate==4.1.1
urllib3==2.3.0
